+++ src/AbstractAdapter.ts
@@ -16,88 +16,31 @@ export abstract class AbstractAdapter implements IStorage {
   // protected type: StorageType;
   protected type: string;
   protected config: AdapterConfig;
-  protected bucketName: string = "";
-  protected initialized: boolean = false;
 
   getType(): string {
     return this.type;
+  }
+
+  public async init(): Promise<ResultObject> {
+    // Except for Backblaze B2 Native API this method doesn't have to do anything
+    // so it doesn't have to be overridden at all, it is only here to make the
+    // API consistent.
+    return Promise.resolve({ error: null, value: "ok" });
   }
 
   public getConfiguration(): AdapterConfig {
     return this.config;
   }
 
-  async test(): Promise<ResultObject> {
-    if (this.initialized === false) {
-      return Promise.resolve({
-        value: null,
-        error: "storage has not been initialized yet; call Storage.init() first",
-      });
-    }
-
-    if (this.bucketName) {
-      let result: ResultObject;
-      try {
-        const { error } = await this.bucketExists(this.bucketName);
-        if (error === null) {
-          result = { value: "ok", error };
-        } else {
-          result = { value: null, error };
-        }
-      } catch (e) {
-        result = {
-          value: null,
-          error: `Looks like the storage configuration is not correct (${e.message})`,
-        };
-      }
-      return Promise.resolve(result);
-    }
-
-    let result: ResultObject;
-    try {
-      const { error } = await this.listBuckets();
-      if (error === null) {
-        result = { value: "ok", error };
-      } else {
-        result = { value: null, error };
-      }
-    } catch (e) {
-      result = {
-        value: null,
-        error: `Looks like the storage configuration is not correct (${e.message})`,
-      };
-    }
-    return Promise.resolve(result);
-  }
-
   async addFileFromPath(params: FilePath): Promise<ResultObject> {
-    if (this.initialized === false) {
-      return Promise.resolve({
-        value: null,
-        error: "storage has not been initialized yet; call Storage.init() first",
-      });
-    }
-
     return await this.store(params);
   }
 
   async addFileFromBuffer(params: FileBuffer): Promise<ResultObject> {
-    if (this.initialized === false) {
-      return Promise.resolve({
-        value: null,
-        error: "storage has not been initialized yet; call Storage.init() first",
-      });
-    }
     return await this.store(params);
   }
 
   async addFileFromReadable(params: FileStream): Promise<ResultObject> {
-    if (this.initialized === false) {
-      return Promise.resolve({
-        value: null,
-        error: "storage has not been initialized yet; call Storage.init() first",
-      });
-    }
     return await this.store(params);
   }
 
@@ -106,8 +49,6 @@ export abstract class AbstractAdapter implements IStorage {
   protected abstract store(param: FilePath): Promise<ResultObject>;
   protected abstract store(param: FileBuffer): Promise<ResultObject>;
   protected abstract store(param: FileStream): Promise<ResultObject>;
-
-  abstract init(): Promise<ResultObject>;
 
   abstract createBucket(name: string, options?: object): Promise<ResultObject>;
 
@@ -118,9 +59,12 @@ export abstract class AbstractAdapter implements IStorage {
   abstract listBuckets(): Promise<ResultObjectBuckets>;
 
   abstract getFileAsReadable(
-    name: string,
+    bucketName: string,
+    fileName: string,
     options?: { start?: number; end?: number }
   ): Promise<ResultObjectReadable>;
+
+  abstract getFileAsURL(): Promise<ResultObject>;
 
   abstract removeFile(bucketName: string, fileName: string): Promise<ResultObject>;