+++ changelog.md
@@ -37,7 +37,8 @@
   };
   ```
 
-- `init` will automatically select (and if necessary create) the bucket if your configuration object or url has a value set for `bucketName`
+- ~~`init` will automatically select (and if necessary create) the bucket if your configuration object or url has a value set for `bucketName`~~
+- Backblaze B2 native API storage requires initial authorization (by calling the async `authorize` function) so `init` will only be implemented for this type of storage. For other storage type `init` will be a stub.
 - The storage instance will no longer hold a reference to the last used or selected bucket in its local state; you will have to provide a bucket name for every bucket operation, for instance `clearBucket`, but also `removeFile`.
 - The storage instance will also no longer hold a reference to all available buckets; a call to `listBuckets` will access the cloud storage service every time it is called; this is handy in case another process or user has created a new bucket.
 - `validateName` will not only perform a local check, it will also check if the name is valid and/or not taken at the cloud storage service.
@@ -52,7 +53,7 @@
 #### test
 
 `test():Promise<string>`<br/>
-`test():Promise<ResultObject>`
+`N/A`
 
 #### selectBucket
 
@@ -217,6 +218,28 @@ type GetFile = {
 }
 
 getFile(GetFile): Promise<ResultObjectStream | ResultObject>
+```
+
+### The `init` function is not required anymore
+
+Only Backblaze B2 Native API storage requires initial authorization by calling the async `authorize` function. So only for this type of storage it is required to call `init` after instantiating and before you call any API method. Although it is implemented (as a stub) in all storage types, for other storage types you don't need to call this method:
+
+```typescript
+const b2 = new Storage("b2://applicationKeyId:applicationKey");
+await b2.init(); // mandatory
+await b2.listBuckets();
+
+const gcs = new Storage("gcs://keyFile.json");
+await gcs.listBuckets();
+```
+
+### The bucket in the config is not automatically selected or created
+
+The bucket name that you've provided with the configuration url or object is available by calling `getConfig`:
+
+```typescript
+const  s3 = new Storage("s3://key:secret@eu-west-2/bucketName");
+await s3.listFiles(s3.getConfig().bucketName, "your-file.jpg')
 ```
 
 # 1.4.7 - 1.5.2